Add triple-slash types reference to trusted-types.
