+++ package.json
@@ -7,10 +7,10 @@
     "commit-amend-build": "scripts/commit-amend-build.sh",
     "prebuild": "rimraf dist/**",
     "dev": "cross-env NODE_ENV=development BABEL_ENV=rollup rollup -w -c -o dist/purify.js",
-    "build": "run-s build:types build:rollup build:fix-cjs-types build:cleanup",
+    "build": "run-s build:types build:rollup build:fix-types build:cleanup",
     "build:types": "tsc --outDir dist/types --declaration --emitDeclarationOnly",
     "build:rollup": "rollup -c",
-    "build:fix-cjs-types": "node ./scripts/fix-cjs-types.js",
+    "build:fix-types": "node ./scripts/fix-types.js",
     "build:umd": "rollup -c -f umd -o dist/purify.js",
     "build:umd:min": "rollup -c -f umd -o dist/purify.min.js -p terser",
     "build:es": "rollup -c -f es -o dist/purify.es.mjs",
@@ -103,6 +103,7 @@
     "@rollup/plugin-replace": "^6.0.1",
     "@rollup/plugin-terser": "^0.4.4",
     "@types/estree": "^1.0.0",
+    "@types/node": "^16.18.120",
     "cross-env": "^7.0.3",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
